refactor(all): reorg, make renderer code-splittable
