+++ src/aurelia-dialog.ts
@@ -22,4 +22,4 @@ export * from './dialog-cancel-error';
 export * from './dialog-result';
 export * from './dialog-service';
 export * from './dialog-controller';
-export { NativeDialogRenderer } from './native-dialog-renderer';
+export { NativeDialogRenderer } from './renderers/native-dialog-renderer';
